2025/5/9——管理端登录、首页以及视频审核前后端联调完成

+++ Y-service/src/main/java/com/yilena/service/entity/po/Log.java
@@ -17,5 +17,5 @@ public class Log {
     private String methodName;
     private String methodParams;
     private String returnValue;
-    private Long costTime;
+    private double costTime;
 }
